chore(asset-registry): reuse Ownable and Initializable to refactor asset registry
