--- packages/asset-registry/contracts/Asset/AssetProducingRegistryLogic.sol
@@ -1,167 +1,0 @@
-pragma solidity ^0.5.0;
-pragma experimental ABIEncoderV2;
-
-import "../Asset/AssetProducingDB.sol";
-import "../AssetContractLookup.sol";
-
-import "../Asset/AssetLogic.sol";
-import "@energyweb/utils-general/contracts/Msc/Owned.sol";
-import "../Interfaces/AssetProducingInterface.sol";
-
-
-/// @title The logic contract for the asset registration
-/// @notice This contract provides the logic that determines how the data is stored
-/// @dev Needs a valid AssetProducingRegistryDB contract to function correctly
-contract AssetProducingRegistryLogic is AssetLogic, AssetProducingInterface {
-
-    event LogNewMeterRead(
-        uint indexed _assetId,
-        uint _oldMeterRead,
-        uint _newMeterRead,
-        uint _timestamp
-    );
-
-    
-    /// @notice Constructor
-    /// @param _userContractLookup usercontract-lookup-contract
-    /// @param _assetContractLookup the asset-lookup-contract
-    constructor(
-        UserContractLookupInterface _userContractLookup, 
-        AssetContractLookupInterface _assetContractLookup
-    ) 
-        RoleManagement(_userContractLookup, address(_assetContractLookup)) 
-        public 
-    {
-    }
-
-	/// @notice Logs meter read
-	/// @param _assetId The id belonging to an entry in the asset registry
-	/// @param _newMeterRead The current meter read of the asset
-	/// @param _lastSmartMeterReadFileHash Last meter read file hash
-    function saveSmartMeterRead(
-        uint _assetId,
-        uint _newMeterRead,
-        string calldata _lastSmartMeterReadFileHash,
-        uint _timestamp
-    )
-        external
-        isInitialized
-    {
-        uint timestamp = _timestamp;
-
-        require(timestamp >= 0, "a timestamp cannot be a negative number");
-        require(timestamp <= block.timestamp + 60, "a timestamp cannot be higher than current block time plus 1 min");
-
-        if (timestamp == 0) {
-            timestamp = block.timestamp;
-        }
-
-        uint createdEnergy = setSmartMeterReadInternal(_assetId, _newMeterRead, _lastSmartMeterReadFileHash, timestamp);
-
-        AssetProducingDB(address(db)).addAssetRead(_assetId, AssetProducingDB.SmartMeterRead({
-            energy: createdEnergy,
-            timestamp: timestamp
-        }));
-    }
-
-	/// @notice creates an asset with the provided parameters
-	/// @param _smartMeter smartmeter of the asset
-	/// @param _owner asset-owner
-	/// @param _active flag if the asset is already active
-	/// @param _propertiesDocumentHash hash of the document with the properties of an asset
-	/// @param _url where to find the documentHash
-	/// @param _numOwnerChanges allowed amount of owner-changes of certificates created by the asset
-	/// @return generated asset-id
-    function createAsset(
-        address _smartMeter,
-        address _owner,
-        bool _active,
-        string calldata _propertiesDocumentHash,
-        string calldata _url,
-        uint _numOwnerChanges
-    )
-        external
-        returns (uint _assetId)
-    {
-        checkBeforeCreation(_owner, _smartMeter);
-
-        AssetGeneral memory a = AssetGeneral({
-            smartMeter: _smartMeter,
-            owner: _owner,
-            lastSmartMeterReadWh: 0,
-            active: true,
-            lastSmartMeterReadFileHash: "",
-            propertiesDocumentHash: _propertiesDocumentHash,
-            url: _url,
-            bundled: false
-        });
-
-        AssetProducingDB.Asset memory _asset = AssetProducingDB.Asset(
-            {assetGeneral: a,
-            maxOwnerChanges: _numOwnerChanges
-            }
-        );
-
-        _assetId =  AssetProducingDB(address(db)).addFullAsset(_asset);
-
-        emit LogAssetCreated(msg.sender, _assetId);
-
-    }
-
-    function getSmartMeterReadsForAsset(uint _assetId)
-        external
-        view
-        returns (
-            AssetProducingDB.SmartMeterRead[] memory reads
-        )
-    {
-        return AssetProducingDB(address(db)).getSmartMeterReadsForAsset(_assetId);
-    }
-
-	/// @notice Gets an asset
-	/// @param _assetId The id belonging to an entry in the asset registry
-	/// @return Full informations of an asset
-    function getAssetById(uint _assetId)
-        external
-        view
-        returns (
-            AssetProducingDB.Asset memory
-        )
-    {
-        return AssetProducingDB(address(db)).getAssetById(_assetId);
-    }
-
-	/// @notice gets an asset by its smartmeter
-	/// @param _smartMeter smartmeter used for by the asset
-	/// @return Asset-Struct
-    function getAssetBySmartMeter(address _smartMeter)
-        external
-        view
-        returns (
-            AssetProducingDB.Asset memory
-        )
-    {
-        return AssetProducingDB(address(db)).getAssetBySmartMeter(_smartMeter);
-    }
-
-	/// @notice checks whether an assets with the provided smartmeter already exists
-	/// @param _smartMeter smartmter of an asset
-	/// @return whether there is already an asset with that smartmeter
-    function checkAssetExist(address _smartMeter) public view returns (bool){
-        return checkAssetGeneralExistingStatus(AssetProducingDB(address(db)).getAssetBySmartMeter(_smartMeter).assetGeneral);
-    }
-
-    function checkAssetExistExternal(address _smartMeter) external view returns (bool) {
-        return checkAssetExist(_smartMeter);
-    }
-
-
-	/// @notice enabes or disables the bundle-functionality
-	/// @param _assetId the id of an asset
-	/// @param _active whether the asset should use bundle functions
-    function setBundleActive(uint _assetId, bool _active) external {
-
-        require(msg.sender == db.getAssetOwner(_assetId),"setBundleActive: not the owner");
-        db.setIsBundled(_assetId, _active);
-    }
-}
