+++ packages/user-registry/package.json
@@ -36,7 +36,7 @@
     "lint-fix": "solium -d contracts --fix && eslint 'src/**/*{.ts,.tsx}' --fix",
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20 -p 8548",
     "start-backend": "PORT=3033 node ../../node_modules/@energyweb/origin-backend/dist/js/src/index.js",
-    "test": "mocha -r ts-node/register src/test/UserContract.test.ts --timeout 60000 --exit",
+    "test": "mocha -r ts-node/register src/test/*.test.ts --timeout 60000 --exit",
     "test:concurrent": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"rm -rf db.sqlite && yarn start-backend\" \"wait-on tcp:8548 && yarn test\"",
     "test:contracts": "yarn test:concurrent",
     "prettier": "prettier --write --config-precedence file-override './src/**/*'",
