+++ packages/user-registry/src/utils/migrateContracts.ts
@@ -1,23 +1,21 @@
 import Web3 from 'web3';
 import { deploy } from '@energyweb/utils-general';
-import { GeneralFunctions } from '../wrappedContracts/GeneralFunctions';
 
+import { UserLogic } from '../wrappedContracts/UserLogic';
 import UserLogicJSON from '../../build/contracts/UserLogic.json';
 
-export async function migrateUserRegistryContracts(web3: Web3, deployKey: string): Promise<any> {
+export async function migrateUserRegistryContracts(
+    web3: Web3,
+    deployKey: string
+): Promise<UserLogic> {
     const privateKeyDeployment = deployKey.startsWith('0x') ? deployKey : `0x${deployKey}`;
 
     const userLogicAddress = (await deploy(web3, UserLogicJSON.bytecode, {
         privateKey: privateKeyDeployment
     })).contractAddress;
 
-    const userLogic = new web3.eth.Contract(UserLogicJSON.abi, userLogicAddress);
+    const userLogic = new UserLogic(web3, userLogicAddress);
+    await userLogic.initialize();
 
-    await new GeneralFunctions(userLogic, web3).send(userLogic.methods.initialize(), {
-        privateKey: privateKeyDeployment
-    });
-
-    return {
-        userLogic: userLogicAddress
-    };
+    return userLogic;
 }
