+++ packages/user-registry/src/wrappedContracts/UserLogic.ts
@@ -22,6 +22,12 @@ export class UserLogic extends GeneralFunctions {
         return this.web3Contract.getPastEvents('allEvents', eventFilter);
     }
 
+    async initialize(txParams?: ISpecialTx) {
+        const method = this.web3Contract.methods.initialize();
+
+        return this.send(method, txParams);
+    }
+
     async deactivateUser(_user: string, txParams?: ISpecialTx) {
         const method = this.web3Contract.methods.deactivateUser(_user);
